rfid-project/services/api/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
influxdb-client==1.38.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic>=2.6.4